# Reduce memory usage of the filecache 