--- a/enterprise/server/remote_execution/filecache/filecache_test.go
+++ b/enterprise/server/remote_execution/filecache/filecache_test.go
@@ -559,7 +559,6 @@ func BenchmarkFilecacheLink(b *testing.B) {
 }
 
 func BenchmarkContainsAdd(b *testing.B) {
-	// Use a simple Claims to speed up filecache.groupIDStringFromContext()
 	ctx := claims.AuthContextFromClaims(context.Background(), &claims.Claims{}, nil)
 	flags.Set(b, "app.log_level", "warn")
 	log.Configure()