Since the LRU now has string keys, we can use those to get the filecache paths, instead of storing the path in the LRU value.

I also simplified the benchmark and made it monitor retained memory usage.

This reduces memory usage by 25% compared to the previous version of the LRU (with hashed keys), and by 33% compared to the current LRU (with string keys). It also speeds it up a bit and reduces allocations.

```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor
                      │ /tmp/baseline │          /tmp/simplekeys           │         /tmp/nopathinvalue         │
                      │    sec/op     │   sec/op     vs base               │   sec/op     vs base               │
ContainsAdd/1K/10G-24     642.8µ ± 2%   616.4µ ± 3%  -4.11% (p=0.000 n=50)   605.5µ ± 2%  -5.80% (p=0.000 n=50)
ContainsAdd/1K/1M-24      14.79m ± 1%   14.96m ± 1%       ~ (p=0.227 n=50)   14.64m ± 2%       ~ (p=0.207 n=50)
geomean                   3.083m        3.037m       -1.50%                  2.977m       -3.43%

                      │ /tmp/baseline │            /tmp/simplekeys            │          /tmp/nopathinvalue           │
                      │ retained-heap │ retained-heap  vs base                │ retained-heap  vs base                │
ContainsAdd/1K/10G-24     314.7k ± 2%     361.6k ± 2%  +14.91% (p=0.000 n=50)     232.9k ± 4%  -25.98% (n=50)
ContainsAdd/1K/1M-24      90.55k ± 6%    101.40k ± 5%  +11.98% (p=0.001 n=50)     69.58k ± 6%  -23.16% (p=0.000 n=50)
geomean                   168.8k          191.5k       +13.44%                    127.3k       -24.58%

                      │ /tmp/baseline │           /tmp/simplekeys           │         /tmp/nopathinvalue          │
                      │     B/op      │     B/op      vs base               │     B/op      vs base               │
ContainsAdd/1K/10G-24    165.3Ki ± 0%   165.2Ki ± 0%  -0.05% (p=0.000 n=50)   165.1Ki ± 0%  -0.12% (n=50)
ContainsAdd/1K/1M-24     1.380Mi ± 0%   1.374Mi ± 0%  -0.45% (p=0.001 n=50)   1.261Mi ± 0%  -8.61% (p=0.000 n=50)
geomean                  483.4Ki        482.2Ki       -0.25%                  461.8Ki       -4.46%

                      │ /tmp/baseline │          /tmp/simplekeys           │     /tmp/nopathinvalue      │
                      │   allocs/op   │  allocs/op   vs base               │  allocs/op   vs base        │
ContainsAdd/1K/10G-24     3.013k ± 0%   3.012k ± 0%  -0.03% (p=0.000 n=50)   3.011k ± 0%   -0.07% (n=50)
ContainsAdd/1K/1M-24      17.27k ± 0%   16.33k ± 0%  -5.42% (n=50)           15.42k ± 0%  -10.70% (n=50)
geomean                   7.213k        7.014k       -2.76%                  6.814k        -5.53%
```